linting

+++ examples/janus/demo/app_januspro.py
@@ -24,18 +24,14 @@ from mindone.utils.seed import set_random_seed
 
 # args
 parser = argparse.ArgumentParser()
-parser.add_argument(
-    "--ms_mode", type=int, default=1, help="mindspore mode, 0: graph, 1: pynative"
-)
+parser.add_argument("--ms_mode", type=int, default=1, help="mindspore mode, 0: graph, 1: pynative")
 parser.add_argument(
     "--model_path",
     type=str,
     default="ckpts/Janus-Pro-7B",
     help="path to model weight folder",
 )
-parser.add_argument(
-    "--share", type=str2bool, default=False, help="private or share demo (public)"
-)
+parser.add_argument("--share", type=str2bool, default=False, help="private or share demo (public)")
 args = parser.parse_args()
 
 # ms init
@@ -48,9 +44,7 @@ if args.ms_mode == 0:
 config = AutoConfig.from_pretrained(args.model_path)
 language_config = config.language_config
 language_config._attn_implementation = "eager"
-vl_gpt = AutoModelForCausalLM.from_pretrained(
-    args.model_path, language_config=language_config, trust_remote_code=True
-)
+vl_gpt = AutoModelForCausalLM.from_pretrained(args.model_path, language_config=language_config, trust_remote_code=True)
 
 vl_gpt = set_model_param_dtype(vl_gpt, ms.bfloat16)
 vl_gpt.set_train(False)
@@ -78,9 +72,9 @@ def multimodal_understanding(image, question, seed, top_p, temperature):
 
     pil_images = [Image.fromarray(image)]
 
-    prepare_inputs = vl_chat_processor(
-        conversations=conversation, images=pil_images, force_batchify=True
-    ).to(ms.bfloat16)
+    prepare_inputs = vl_chat_processor(conversations=conversation, images=pil_images, force_batchify=True).to(
+        ms.bfloat16
+    )
 
     inputs_embeds = vl_gpt.prepare_inputs_embeds(**prepare_inputs)
 
@@ -122,12 +116,8 @@ def generate(
         if i % 2 != 0:
             tokens[i, 1:-1] = vl_chat_processor.pad_id
 
-    inputs_embeds = vl_gpt.language_model.get_input_embeddings()(tokens).to(
-        vl_gpt.dtype
-    )
-    generated_tokens = mint.zeros(
-        (parallel_size, image_token_num_per_image), dtype=ms.int32
-    )
+    inputs_embeds = vl_gpt.language_model.get_input_embeddings()(tokens).to(vl_gpt.dtype)
+    generated_tokens = mint.zeros((parallel_size, image_token_num_per_image), dtype=ms.int32)
 
     use_cache = False
     outputs = None
@@ -152,18 +142,14 @@ def generate(
             next_token = mint.argmax(logits, dim=-1, keepdim=True)
 
         generated_tokens[:, i] = next_token.squeeze(axis=-1)
-        next_token = mint.cat(
-            [next_token.unsqueeze(dim=1), next_token.unsqueeze(dim=1)], dim=1
-        ).view(-1)
+        next_token = mint.cat([next_token.unsqueeze(dim=1), next_token.unsqueeze(dim=1)], dim=1).view(-1)
 
         img_embeds = vl_gpt.prepare_gen_img_embeds(next_token)
 
         if use_cache:
             inputs_embeds = img_embeds.unsqueeze(dim=1)
         else:
-            inputs_embeds = ops.concat(
-                (inputs_embeds, img_embeds.unsqueeze(dim=1)), axis=1
-            )
+            inputs_embeds = ops.concat((inputs_embeds, img_embeds.unsqueeze(dim=1)), axis=1)
 
     patches = vl_gpt.gen_vision_model.decode_code(
         generated_tokens.to(dtype=ms.int32),
@@ -216,9 +202,7 @@ def generate_image(prompt, seed=None, guidance=5, t2i_temperature=1.0):
         parallel_size=parallel_size,
         temperature=t2i_temperature,
     )
-    images = unpack(
-        patches, width // 16 * 16, height // 16 * 16, parallel_size=parallel_size
-    )
+    images = unpack(patches, width // 16 * 16, height // 16 * 16, parallel_size=parallel_size)
 
     # return [Image.fromarray(images[i]).resize((768, 768), Image.LANCZOS) for i in range(parallel_size)]
     return [Image.fromarray(images[i]) for i in range(parallel_size)]
@@ -232,12 +216,8 @@ with gr.Blocks() as demo:
         with gr.Column():
             question_input = gr.Textbox(label="Question")
             und_seed_input = gr.Number(label="Seed", precision=0, value=42)
-            top_p = gr.Slider(
-                minimum=0, maximum=1, value=0.95, step=0.05, label="top_p"
-            )
-            temperature = gr.Slider(
-                minimum=0, maximum=1, value=0.1, step=0.05, label="temperature"
-            )
+            top_p = gr.Slider(minimum=0, maximum=1, value=0.95, step=0.05, label="top_p")
+            temperature = gr.Slider(minimum=0, maximum=1, value=0.1, step=0.05, label="temperature")
 
     understanding_button = gr.Button("Chat")
     understanding_output = gr.Textbox(label="Response")
@@ -260,16 +240,10 @@ with gr.Blocks() as demo:
     gr.Markdown(value="# Text-to-Image Generation")
 
     with gr.Row():
-        cfg_weight_input = gr.Slider(
-            minimum=1, maximum=10, value=5, step=0.5, label="CFG Weight"
-        )
-        t2i_temperature = gr.Slider(
-            minimum=0, maximum=1, value=1.0, step=0.05, label="temperature"
-        )
+        cfg_weight_input = gr.Slider(minimum=1, maximum=10, value=5, step=0.5, label="CFG Weight")
+        t2i_temperature = gr.Slider(minimum=0, maximum=1, value=1.0, step=0.05, label="temperature")
 
-    prompt_input = gr.Textbox(
-        label="Prompt. (Prompt in more detail can help produce better images!)"
-    )
+    prompt_input = gr.Textbox(label="Prompt. (Prompt in more detail can help produce better images!)")
     seed_input = gr.Number(label="Seed (Optional)", precision=0, value=12345)
 
     generation_button = gr.Button("Generate Images")
@@ -316,6 +290,4 @@ with gr.Blocks() as demo:
 if args.share:
     demo.launch(share=True)
 else:
-    demo.queue(concurrency_count=1, max_size=10).launch(
-        server_name="127.0.0.1", server_port=37906, root_path="/path"
-    )
+    demo.queue(concurrency_count=1, max_size=10).launch(server_name="127.0.0.1", server_port=37906, root_path="/path")
